Split UUID from package object so coverage can track it
